dns_server_requirements.txt
dnslib
BeautifulSoup4
dnspython
netaddr
watchdog
